+++ plugins/src/main/scala/com/stratio/sparta/plugin/workflow/output/jdbc/JdbcOutputStep.scala
@@ -31,6 +31,7 @@ class JdbcOutputStep(name: String, xDSession: XDSession, properties: Map[String,
   lazy val url = properties.getString("url")
   lazy val tlsEnable = Try(properties.getBoolean("tlsEnabled")).getOrElse(false)
   lazy val jdbcSaveMode = TransactionTypes.withName(properties.getString("jdbcSaveMode", "STATEMENT"))
+  lazy val isCaseSensitive = Try(properties.getBoolean("caseSensitiveEnabled")).getOrElse(true)
   lazy val failFast = Try(properties.getBoolean("failFast")).getOrElse(false)
   val sparkConf = xDSession.conf.getAll
   val securityUri = getDataStoreUri(sparkConf)
@@ -69,8 +70,9 @@ class JdbcOutputStep(name: String, xDSession: XDSession, properties: Map[String,
     if (dataFrame.schema.fields.nonEmpty) {
       val tableName = getTableNameFromOptions(options)
       val sparkSaveMode = getSparkSaveMode(saveMode)
-      val connectionProperties = new JDBCOptions(urlWithSSL,
-        tableName,
+      val connectionProperties = new JDBCOptions(
+        urlWithSSL,
+        if (isCaseSensitive) quoteTable(tableName) else tableName,
         propertiesWithCustom.mapValues(_.toString).filter(_._2.nonEmpty)
       )