+++ serving-core/src/test/scala/com/stratio/sparta/serving/core/models/workflow/WorkflowValidationTest.scala
@@ -34,7 +34,7 @@ class WorkflowValidationTest extends WordSpec with Matchers with MockitoSugar {
   val settingsModel = Settings(
     GlobalSettings(executionMode = WorkflowExecutionMode.local),
     StreamingSettings(
-      JsoneyString("6s"), None, None, None, None, None, None, CheckpointSettings(JsoneyString("test/test"))),
+      JsoneyString("6s"), None, None, None, None, None, None, None, CheckpointSettings(JsoneyString("test/test"))),
     SparkSettings(
       JsoneyString("local[*]"), sparkKerberos = false, sparkDataStoreTls = false, sparkMesosSecurity = false,
       None, SubmitArguments(), SparkConf(SparkResourcesConf())
@@ -417,7 +417,7 @@ class WorkflowValidationTest extends WordSpec with Matchers with MockitoSugar {
       val wrongSettingsModel = Settings(
         GlobalSettings(executionMode = WorkflowExecutionMode.marathon),
         StreamingSettings(
-          JsoneyString("6s"), None, None, None, None, None, None, CheckpointSettings(JsoneyString("test/test"))),
+          JsoneyString("6s"), None, None, None, None, None, None, None, CheckpointSettings(JsoneyString("test/test"))),
         SparkSettings(
           JsoneyString("mesos://zk://master.mesos:2181/mesos"), sparkKerberos = false, sparkDataStoreTls = false,
           sparkMesosSecurity = false, None, SubmitArguments(deployMode = Option(DeployMode.cluster)),
@@ -434,7 +434,7 @@ class WorkflowValidationTest extends WordSpec with Matchers with MockitoSugar {
       val wrongSettingsModel = Settings(
         GlobalSettings(executionMode = WorkflowExecutionMode.marathon),
         StreamingSettings(
-          JsoneyString("6s"), None, None, None, None, None, None, CheckpointSettings(JsoneyString("test/test"))),
+          JsoneyString("6s"), None, None, None, None, None, None, None, CheckpointSettings(JsoneyString("test/test"))),
         SparkSettings(
           JsoneyString("mesos://zk://master.mesos:2181/mesos"), sparkKerberos = false, sparkDataStoreTls = false,
           sparkMesosSecurity = false, None, SubmitArguments(),
@@ -451,7 +451,7 @@ class WorkflowValidationTest extends WordSpec with Matchers with MockitoSugar {
       val wrongSettingsModel = Settings(
         GlobalSettings(executionMode = WorkflowExecutionMode.marathon),
         StreamingSettings(
-          JsoneyString("6s"), None, None, None, None, None, None, CheckpointSettings(JsoneyString("test/test"))),
+          JsoneyString("6s"), None, None, None, None, None, None, None, CheckpointSettings(JsoneyString("test/test"))),
         SparkSettings(
           JsoneyString("mesos://zk://master.mesos:2181/mesos"), sparkKerberos = false, sparkDataStoreTls = false,
           sparkMesosSecurity = false, None, SubmitArguments(),
@@ -468,7 +468,7 @@ class WorkflowValidationTest extends WordSpec with Matchers with MockitoSugar {
       val wrongSettingsModel = Settings(
         GlobalSettings(executionMode = WorkflowExecutionMode.marathon),
         StreamingSettings(
-          JsoneyString("6s"), None, None, None, None, None, None, CheckpointSettings(JsoneyString("test/test"))),
+          JsoneyString("6s"), None, None, None, None, None, None, None, CheckpointSettings(JsoneyString("test/test"))),
         SparkSettings(
           JsoneyString("mesos://zk://master.mesos:2181/mesos"), sparkKerberos = false, sparkDataStoreTls = false,
           sparkMesosSecurity = false, None, SubmitArguments(),
@@ -485,7 +485,7 @@ class WorkflowValidationTest extends WordSpec with Matchers with MockitoSugar {
       val wrongSettingsModel = Settings(
         GlobalSettings(executionMode = WorkflowExecutionMode.marathon),
         StreamingSettings(
-          JsoneyString("6s"), None, None, None, None, None, None, CheckpointSettings(JsoneyString("test/test"))),
+          JsoneyString("6s"), None, None, None, None, None, None, None, CheckpointSettings(JsoneyString("test/test"))),
         SparkSettings(
           JsoneyString("mesos://zk://master.mesos:2181/mesos"), sparkKerberos = false, sparkDataStoreTls = false,
           sparkMesosSecurity = false, None, SubmitArguments(),
@@ -502,7 +502,7 @@ class WorkflowValidationTest extends WordSpec with Matchers with MockitoSugar {
       val wrongSettingsModel = Settings(
         GlobalSettings(executionMode = WorkflowExecutionMode.marathon),
         StreamingSettings(
-          JsoneyString("6s"), None, None, None, None, None, None, CheckpointSettings(JsoneyString("test/test"))),
+          JsoneyString("6s"), None, None, None, None, None, None, None, CheckpointSettings(JsoneyString("test/test"))),
         SparkSettings(
           JsoneyString("mesos://zk://master.mesos:2181/mesos"), sparkKerberos = false, sparkDataStoreTls = false,
           sparkMesosSecurity = false, None, SubmitArguments(),
@@ -519,7 +519,7 @@ class WorkflowValidationTest extends WordSpec with Matchers with MockitoSugar {
       val wrongSettingsModel = Settings(
         GlobalSettings(executionMode = WorkflowExecutionMode.marathon),
         StreamingSettings(
-          JsoneyString("6s"), None, None, None, None, None, None, CheckpointSettings(JsoneyString("test/test"))),
+          JsoneyString("6s"), None, None, None, None, None, None, None, CheckpointSettings(JsoneyString("test/test"))),
         SparkSettings(
           JsoneyString("mesos://zk://master.mesos:2181/mesos"), sparkKerberos = false, sparkDataStoreTls = false,
           sparkMesosSecurity = false, None, SubmitArguments(),
@@ -536,7 +536,7 @@ class WorkflowValidationTest extends WordSpec with Matchers with MockitoSugar {
       val wrongSettingsModel = Settings(
         GlobalSettings(executionMode = WorkflowExecutionMode.marathon),
         StreamingSettings(
-          JsoneyString("6s"), None, None, None, None, None, None, CheckpointSettings(JsoneyString("test/test"))),
+          JsoneyString("6s"), None, None, None, None, None, None, None, CheckpointSettings(JsoneyString("test/test"))),
         SparkSettings(
           JsoneyString("mesos://zk://master.mesos:2181/mesos"), sparkKerberos = false, sparkDataStoreTls = false,
           sparkMesosSecurity = false, None, SubmitArguments(),
@@ -557,7 +557,7 @@ class WorkflowValidationTest extends WordSpec with Matchers with MockitoSugar {
       val wrongSettingsModel = Settings(
         GlobalSettings(executionMode = WorkflowExecutionMode.marathon),
         StreamingSettings(
-          JsoneyString("6s"), None, None, None, None, None, None, CheckpointSettings(JsoneyString("test/test"), false)),
+          JsoneyString("6s"), None, None, None, None, None, None, None, CheckpointSettings(JsoneyString("test/test"), false)),
         SparkSettings(
           JsoneyString("mesos://zk://master.mesos:2181/mesos"), sparkKerberos = false, sparkDataStoreTls = false,
           sparkMesosSecurity = false, None, SubmitArguments(),